requirements.txt
# Core dependencies
numpy>=1.19.0
pandas>=1.3.0
tqdm>=4.62.0
psutil>=5.9.0

# Web scraping
requests>=2.26.0
beautifulsoup4>=4.10.0

# Semantic analysis
sentence-transformers>=2.2.0
faiss-cpu>=1.7.0
scikit-learn>=1.0.0
torch>=1.11.0

# Visualization and dashboard
streamlit>=1.10.0
plotly>=5.5.0

# Utilities
python-dateutil>=2.8.2
pytz>=2021.3

# Testen
pytest>=7.0.0
coverage>=6.0.0

# Linting
flake8>=5.0.0

# Optionale Abhängigkeiten
pygame>=2.1.0  # Für Soundausgabe im Cursor-Monitor 